backend/requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
python-jose[cryptography]==3.2.0
passlib==1.7.4
python-multipart==0.0.5
pymongo==3.12.0
python-dotenv==0.19.0
bcrypt==3.2.0
pydantic==1.8.2
email-validator==1.1.3
motor==2.5.0 